# Conflicts:
#	app.js
#	db/migrations/3-store.js
#	db/migrations/4-menu-category.js
#	db/models/store.js
#	db/models/user.js
#	package.json

+++ db/migrations/5-menu.js
@@ -7,7 +7,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.SMALLINT.UNSIGNED,
+        type: Sequelize.MEDIUMINT.UNSIGNED,
       },
       storeId: {
         allowNull: false,
